+++ visualization/app/codeCharta/ui/codeMap/threeViewer/threeSceneService.ts
@@ -4,8 +4,22 @@ import { Group } from "three"
 import { CodeMapMesh } from "../rendering/codeMapMesh"
 import { CodeMapBuilding } from "../rendering/codeMapBuilding"
 import { SettingsService } from "../../../state/settingsService/settings.service"
+import { CodeMapPreRenderServiceSubscriber, CodeMapPreRenderService } from "../codeMap.preRender.service"
+import { CodeMapNode } from "../../../codeCharta.model"
+import { IRootScopeService } from "angular"
 
-export class ThreeSceneService {
+export interface BuildingSelectedEventSubscriber {
+	onBuildingSelected(selectedBuilding: CodeMapBuilding)
+}
+
+export interface BuildingDeselectedEventSubscriber {
+	onBuildingDeselected()
+}
+
+export class ThreeSceneService implements CodeMapPreRenderServiceSubscriber {
+	private static readonly BUILDING_SELECTED_EVENT = "building-selected"
+	private static readonly BUILDING_DESELECTED_EVENT = "building-deselected"
+
 	public scene: Scene
 	public labels: Group
 	public edgeArrows: Group
@@ -17,9 +31,10 @@ export class ThreeSceneService {
 	private highlighted: CodeMapBuilding = null
 	private highlightedBuildings: CodeMapBuilding[] = []
 
-	constructor(private settingsService: SettingsService) {
-		this.scene = new THREE.Scene()
+	constructor(private $rootScope: IRootScopeService, private settingsService: SettingsService) {
+		CodeMapPreRenderService.subscribe(this.$rootScope, this)
 
+		this.scene = new THREE.Scene()
 		this.mapGeometry = new THREE.Group()
 		this.lights = new THREE.Group()
 		this.labels = new THREE.Group()
@@ -31,6 +46,10 @@ export class ThreeSceneService {
 		this.scene.add(this.edgeArrows)
 		this.scene.add(this.labels)
 		this.scene.add(this.lights)
+	}
+
+	public onRenderMapChanged(map: CodeMapNode) {
+		this.reselectBuilding()
 	}
 
 	public highlightBuilding(building: CodeMapBuilding) {
@@ -56,11 +75,13 @@ export class ThreeSceneService {
 		const color = this.settingsService.getSettings().appSettings.mapColors.selected
 		this.getMapMesh().selectBuilding(building, this.selected, color)
 		this.selected = building
+		this.$rootScope.$broadcast(ThreeSceneService.BUILDING_SELECTED_EVENT, this.selected)
 	}
 
 	public clearSelection() {
 		if (this.selected) {
 			this.getMapMesh().clearSelection(this.selected)
+			this.$rootScope.$broadcast(ThreeSceneService.BUILDING_DESELECTED_EVENT)
 		}
 		if (this.highlighted) {
 			this.getMapMesh().highlightBuilding(this.highlighted, null, this.settingsService.getSettings())
@@ -125,5 +146,26 @@ export class ThreeSceneService {
 
 	public getHighlightedBuilding(): CodeMapBuilding {
 		return this.highlighted
+	}
+
+	private reselectBuilding() {
+		if (this.selected) {
+			const buildingToSelect: CodeMapBuilding = this.getMapMesh().getBuildingByPath(this.selected.node.path)
+			if (buildingToSelect) {
+				this.selectBuilding(buildingToSelect)
+			}
+		}
+	}
+
+	public static subscribeToBuildingDeselectedEvents($rootScope: IRootScopeService, subscriber: BuildingDeselectedEventSubscriber) {
+		$rootScope.$on(this.BUILDING_DESELECTED_EVENT, e => {
+			subscriber.onBuildingDeselected()
+		})
+	}
+
+	public static subscribeToBuildingSelectedEvents($rootScope: IRootScopeService, subscriber: BuildingSelectedEventSubscriber) {
+		$rootScope.$on(this.BUILDING_SELECTED_EVENT, (e, selectedBuilding: CodeMapBuilding) => {
+			subscriber.onBuildingSelected(selectedBuilding)
+		})
 	}
 }